pyproject.toml
[tool.poetry]
name = "bili-danmu"
version = "0.2.5"
description = "A modern library for receiving danmu from bilibili livestream"
authors = ["WorldObservationLog <wolc@duck.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "danmu"}]


[tool.poetry.dependencies]
python = "^3.10"
websockets = "^10.4"
brotlipy = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.urls]
"GitHub" = "https://github.com/WorldObservationLog/bili-danmu"
